hoist out get_disk_usage
