--- src/app/data_harvester/disks/system/macos/mod.rs
@@ -1,4 +1,0 @@
-mod io_counters;
-pub(crate) use io_counters::*;
-
-mod io_kit;
