+++ src/app/data_harvester/disks/unix/linux/mod.rs
@@ -2,4 +2,4 @@ mod partition;
 pub(crate) use partition::*;
 
 mod io_counters;
-pub(crate) use io_counters::*;
+pub use io_counters::*;
